feat: expose workflow set function to be used from other scripts

+++ Configs/.local/lib/hyde/workflows.sh
@@ -24,6 +24,7 @@ Usage: $0 [OPTIONS]
 
 Options:
     --select | -S       Select a workflow from the available options
+    --set               Set the given workflow
     --waybar            Get workflow info for Waybar
     --help   | -h       Show this help message
     
@@ -36,7 +37,7 @@ if [ -z "${*}" ]; then
 fi
 
 # Define long options
-LONG_OPTS="select,waybar,help"
+LONG_OPTS="select,set:,waybar,help"
 SHORT_OPTS="Sh"
 # Parse options
 PARSED=$(getopt --options ${SHORT_OPTS} --longoptions "${LONG_OPTS}" --name "$0" -- "$@")
@@ -172,6 +173,20 @@ while true; do
   case "$1" in
   -S | --select)
     fn_select
+    # refresh waybar module only if waybar is running
+    if pgrep -x waybar >/dev/null; then
+      pkill -RTMIN+7 waybar
+    fi
+
+    exit 0
+    ;;
+  --set)
+    if [ -z "$2" ]; then
+      echo "Error: --set requires a workflow name"
+      exit 1
+    fi
+    set_conf "HYPR_WORKFLOW" "$2"
+    fn_update
     # refresh waybar module only if waybar is running
     if pgrep -x waybar >/dev/null; then
       pkill -RTMIN+7 waybar
